problem-solver/py/requirements.txt
py-sc-kpm==0.3.0
py-sc-client==0.3.1
requests==2.31.0
SpeechRecognition